move updateStatus functions into separate file
